Refactor: Update import paths for Photographer and Studio types in FormCreateStudio components
