+++ src/components/pages/FormCreateStudio/SectionImages.tsx
@@ -4,9 +4,7 @@ import FormLabel from "@mui/joy/FormLabel";
 import Button from "@mui/joy/Button";
 import SvgIcon from "@mui/joy/SvgIcon";
 import { styled } from "@mui/joy";
-import { selectStudio } from "@/selectors/studioSelector";
-import { useSelector } from "react-redux";
-import { PortfolioPhoto, Studio } from "@/types/types";
+import { PortfolioPhoto } from "@/types/studio";
 import React from "react";
 
 interface SectionImagesProps {
